Add layout override to OdinGraph
The OdinGraph (which instantiates a Plotly graph) previously set
the layout directly. This modification allows the user to pass in an
optional override for the layout, which can be useful for altering the
presentation and scaling of axes.

Some more defaults have been added to the layout for heatmaps so that
they are plotted with a square 1:1 aspect by default.